+++ python/models/compras.py
@@ -48,7 +48,7 @@ class Gastos(db.Model,BaseMixin,AuditMixin):
 
     @validates('importe')
     def validate_non_negative(self, key, value):
-        if Decimal(value) < 0:
+        if float(value) < 0:
             raise ValueError(f"{key.replace('_',' ').capitalize()} no puede ser negativo")
         return value
 
@@ -68,7 +68,7 @@ class GastosRecurrentes(db.Model,BaseMixin,AuditMixin):
 
     @validates('importe')
     def validate_non_negative(self, key, value):
-        if Decimal(value) < 0:
+        if float(value) < 0:
             raise ValueError(f"{key.replace('_',' ').capitalize()} no puede ser negativo")
         return value
 
@@ -85,7 +85,7 @@ class PreciosDeProveedores(db.Model,BaseMixin,AuditMixin):
 
     @validates('precio_unitario')
     def validate_non_negative(self, key, value):
-        if Decimal(value) < 0:
+        if float(value) < 0:
             raise ValueError(f"{key.replace('_',' ').capitalize()} no puede ser negativo")
         return value
